refactor(webapp): move DevToolPanel to features/, rewrite AGENTS.md
DevToolPanel has its own Zustand store (store/feature-flags.ts), making
it inconsistent with Playground — both belong in features/ by the same
rule. Move it and update the three import sites.

AGENTS.md rewrite:
- Consistent directory paths (no src/ prefix in table)
- Remove all phase-specific wording (Phase 2A etc.)
- "app chrome" -> "app shell"
- Extend PascalCase convention to all React component files, not just ui/
- Remove migration context section
- Clarify app/ as bootstrap-only (no independent state)
