+++ packages/webapp/src/layout/AppSidebar/ProfileDropdown.tsx
@@ -2,7 +2,7 @@ import { ChevronsUpDown, CreditCard, LogOut, SlidersHorizontal, Sparkle, UserRou
 import { useMemo } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-import { isDevToolsEnabled, useDevPanelStore } from '@/app/DevToolPanel';
+import { isDevToolsEnabled, useDevPanelStore } from '@/features/DevToolPanel';
 import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuSeparator, DropdownMenuTrigger } from '@/components-v2/ui/DropdownMenu';
 import { SidebarMenu, SidebarMenuItem } from '@/components-v2/ui/Sidebar';
 import { useMeta } from '@/hooks/useMeta';
